requirements.txt
python-vlc
PyQt5
mysql-connector-python